@@ -3,20 +3,23 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@material-table/core": "^4.3.36",
-    "@material-ui/core": "^4.12.3",
-    "@material-ui/icons": "^4.11.2",
-    "axios": "^0.24.0",
+    "@emotion/react": "^11.10.6",
+    "@emotion/styled": "^11.10.6",
+    "@fontsource/roboto": "^4.5.8",
+    "@mui/icons-material": "^5.11.9",
+    "@mui/material": "^5.11.10",
+    "@mui/x-data-grid": "^5.17.25",
+    "axios": "^1.3.4",
     "mapbox-gl": "npm:empty-npm-package@^1.0.0",
-    "maplibre-gl": "^2.1.7",
-    "markdown-to-jsx": "^7.1.7",
+    "maplibre-gl": "^2.4.0",
+    "markdown-to-jsx": "^7.1.9",
     "moment": "^2.29.4",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-map-gl": "^7.0.9",
-    "react-router-dom": "^5.3.0",
-    "react-scripts": "^5.0.0",
-    "typeface-roboto": "^1.1.13"
+    "react": "^18.2.0",
+    "react-carbonbadge": "^2.0.0",
+    "react-dom": "^18.2.0",
+    "react-map-gl": "^7.0.21",
+    "react-router-dom": "^6.8.1",
+    "react-scripts": "^5.0.1"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -48,7 +51,7 @@
     ]
   },
   "devDependencies": {
-    "gh-pages": "^3.2.1",
+    "gh-pages": "^3.2.3",
     "standard": "^16.0.4"
   }
 }
