@@ -1,23 +1,17 @@
 import React, { useState, useEffect, useRef } from 'react'
-import { withRouter } from 'react-router'
+import { useNavigate } from 'react-router-dom';
 
-import Box from '@material-ui/core/Box'
-import CircularProgress from '@material-ui/core/CircularProgress'
-import Divider from '@material-ui/core/Divider'
-import IconButton from '@material-ui/core/IconButton'
-import InputBase from '@material-ui/core/InputBase'
-import ListSubheader from '@material-ui/core/ListSubheader'
-import Menu from '@material-ui/core/Menu'
-import MenuItem from '@material-ui/core/MenuItem'
-import Tooltip from '@material-ui/core/Tooltip'
+import Box from '@mui/material/Box'
+import CircularProgress from '@mui/material/CircularProgress'
+import IconButton from '@mui/material/IconButton'
+import InputBase from '@mui/material/InputBase'
+import Tooltip from '@mui/material/Tooltip'
 
-import { fade } from '@material-ui/core/styles/colorManipulator'
-import { makeStyles } from '@material-ui/core/styles'
+import { alpha } from '@mui/material/styles';
 
-import ClearIcon from '@material-ui/icons/ClearTwoTone'
-import MyLocationIcon from '@material-ui/icons/MyLocationTwoTone'
-import SearchIcon from '@material-ui/icons/SearchTwoTone'
-import SettingsIcon from '@material-ui/icons/SettingsTwoTone'
+import ClearIcon from '@mui/icons-material/ClearTwoTone'
+import MyLocationIcon from '@mui/icons-material/MyLocationTwoTone'
+import SearchIcon from '@mui/icons-material/SearchTwoTone'
 
 import { useSearchStateValue } from './context/searchState'
 import { useViewStateValue } from './context/viewState'
@@ -25,40 +19,7 @@ import { useViewStateValue } from './context/viewState'
 import * as geoHelper from './helpers/geo'
 import * as urlHelper from './helpers/url'
 
-const useStyles = makeStyles((theme) => ({
-  grow: {
-    flexGrow: 1
-  },
-  iconButton: {
-    padding: theme.spacing()
-  },
-  iconProgress: {
-    margin: theme.spacing()
-  },
-  inputInput: {
-    paddingTop: theme.spacing(),
-    paddingRight: theme.spacing(),
-    paddingBottom: theme.spacing(),
-    paddingLeft: theme.spacing(2),
-    fontWeight: theme.typography.fontWeightBold
-  },
-  search: {
-    position: 'relative',
-    border: '1px solid #E0E0E0',
-    borderColor: theme.palette.outline.main,
-    borderRadius: theme.shape.borderRadius,
-    backgroundColor: fade(theme.palette.common.white, 0.8),
-    '&:hover': {
-      backgroundColor: fade(theme.palette.common.white, 0.9)
-    },
-    marginLeft: 0,
-    marginRight: theme.spacing(),
-    display: 'flex',
-    maxWidth: 240
-  }
-}))
-
-function usePrevious (value) {
+function usePrevious(value) {
   const ref = useRef()
   useEffect(() => {
     ref.current = value
@@ -66,24 +27,30 @@ function usePrevious (value) {
   return ref.current
 }
 
-function PostcodeSearch (props) {
-  const { settings } = props
-  const [{ searchType, searchPostcode, searchPosition }, dispatchSearch] = useSearchStateValue() //eslint-disable-line
-  const [{ loadingPostcode, loadingLocation }, dispatchView] = useViewStateValue() //eslint-disable-line
+function PostcodeSearch(props) {
+  const [{ searchType, searchPostcode, searchPosition }, dispatchSearch] =
+    useSearchStateValue() //eslint-disable-line
+  const [{ loadingPostcode, loadingLocation }, dispatchView] =
+    useViewStateValue() //eslint-disable-line
 
   const [tempPostcode, setTempPostcode] = useState(searchPostcode || '')
-  const [anchor, setAnchor] = useState(null)
 
   const prevProps = usePrevious({ searchPostcode })
 
+  let navigate = useNavigate();
+
   useEffect(() => {
-    if (prevProps && searchPostcode !== prevProps.searchPostcode) setTempPostcode(searchPostcode)
+    if (prevProps && searchPostcode !== prevProps.searchPostcode)
+      setTempPostcode(searchPostcode)
   }, [searchPostcode, prevProps])
 
   const getLocation = async () => {
     if (!loadingLocation) {
       dispatchView({ type: 'ToggleLoadingLocation' })
-      const pos = (searchPosition.length > 0 ? searchPosition : (await geoHelper.getCurrentPosition()))
+      const pos =
+        searchPosition.length > 0
+          ? searchPosition
+          : await geoHelper.getCurrentPosition()
       dispatchSearch({ type: 'SetLocation', searchPosition: pos })
       const postcode = await getLocationPostcode(pos)
       dispatchView({ type: 'ToggleLoadingLocation' })
@@ -91,7 +58,7 @@ function PostcodeSearch (props) {
     }
   }
 
-  const getLocationPostcode = async (location) => {
+  const getLocationPostcode = async location => {
     if (location.length > 0) {
       const postcode = await geoHelper.getCurrentPostcode(...location)
       setTempPostcode(postcode)
@@ -99,25 +66,19 @@ function PostcodeSearch (props) {
     }
   }
 
-  const openSettingsMenu = (e) => setAnchor(e.currentTarget)
-
-  const closeSettingsMenu = () => setAnchor(null)
-
-  const setSearchDistance = (searchDistance) => {
-    closeSettingsMenu()
-    dispatchSearch({ type: 'SetSearchDistance', searchDistance: searchDistance })
-    if (searchType === 'postcode') postcodeSearch(tempPostcode)
-  }
-
   const clearSearch = () => {
     setTempPostcode('')
     dispatchSearch({ type: 'ClearAll' })
-    urlHelper.clearService(props.history)
+    urlHelper.clearService(navigate)
   }
 
   const postcodeSearch = async (postcode = tempPostcode) => {
     if (!postcode || postcode === '') {
-      dispatchView({ type: 'ShowNotification', notificationMessage: 'Please enter a postcode before searching', notificationSeverity: 'warning' })
+      dispatchView({
+        type: 'ShowNotification',
+        notificationMessage: 'Please enter a postcode before searching',
+        notificationSeverity: 'warning'
+      })
       return
     }
     dispatchView({ type: 'ToggleLoadingPostcode' })
@@ -126,102 +87,117 @@ function PostcodeSearch (props) {
       clearSearch()
       const service = await geoHelper.getPostcode(postcode.trim())
       if (service && service.location && service.location.length > 0) {
-        dispatchSearch({ type: 'SetPostcodeSearch', searchPostcode: postcode, searchPosition: service.location })
-        dispatchView({ type: 'SetMapPosition', mapPosition: service.location, mapZoom: 14 })
+        dispatchSearch({
+          type: 'SetPostcodeSearch',
+          searchPostcode: postcode,
+          searchPosition: service.location
+        })
+        dispatchView({
+          type: 'SetMapPosition',
+          mapPosition: service.location,
+          mapZoom: 14
+        })
       } else {
-        dispatchView({ type: 'ShowNotification', notificationMessage: 'We could not find that postcode', notificationSeverity: 'error' })
+        dispatchView({
+          type: 'ShowNotification',
+          notificationMessage: 'We could not find that postcode',
+          notificationSeverity: 'error'
+        })
       }
     } else {
-      dispatchView({ type: 'ShowNotification', notificationMessage: 'We could not find that postcode', notificationSeverity: 'error' })
+      dispatchView({
+        type: 'ShowNotification',
+        notificationMessage: 'We could not find that postcode',
+        notificationSeverity: 'error'
+      })
     }
     dispatchView({ type: 'ToggleLoadingPostcode' })
   }
 
-  const classes = useStyles()
-
   return (
-    <div className={classes.search}>
+    <Box sx={{
+      position: 'relative',
+      border: 1,
+      borderColor: '#ccc',
+      borderRadius: 1,
+      backgroundColor: (theme) => alpha(theme.palette.secondary.main, 0.10),
+      '&:hover': {
+        backgroundColor: (theme) => alpha(theme.palette.secondary.main, 0.20),
+      },
+      marginLeft: 0,
+      paddingLeft: 0,
+      whitespace: 'nowrap',
+      display: 'inline-flex',
+    }}>
       <InputBase
         placeholder='Postcode'
-        classes={{
-          input: classes.inputInput
-        }}
         value={tempPostcode}
-        onChange={(e) => setTempPostcode(e.target.value.toUpperCase())}
-        onKeyDown={(e) => { if (e.keyCode === 13) postcodeSearch() }}
+        onChange={e => setTempPostcode(e.target.value.toUpperCase())}
+        onKeyDown={e => {
+          if (e.key === 'Enter') postcodeSearch()
+        }}
         inputProps={{ 'aria-label': 'search by postcode' }}
+        sx={{
+          padding: theme => theme.spacing(1),
+          maxWidth: 110,
+        }}
       />
-      {searchType === 'postcode'
-        ? (
-          <Tooltip title='Clear search'>
-            <IconButton
-              aria-label='Clear search'
-              className={classes.iconButton}
-              onClick={() => clearSearch()}
-            >
-              <ClearIcon />
-            </IconButton>
-          </Tooltip>
-        )
-        : null}
-      <div className={classes.grow} />
+      {searchType === 'postcode' ? (
+        <Tooltip title='Clear search'>
+          <IconButton
+            aria-label='Clear search'
+            onClick={() => clearSearch()}
+            size="large">
+            <ClearIcon />
+          </IconButton>
+        </Tooltip>
+      ) : null}
       <Tooltip title='Search by postcode'>
-        {!loadingPostcode
-          ? (
-            <IconButton
-              aria-label='Search'
-              color='primary'
-              className={classes.iconButton}
-              onClick={() => postcodeSearch()}
-            >
-              <SearchIcon />
-            </IconButton>
-          ) : <Box position='relative' display='inline-flex' className={classes.iconProgress}><CircularProgress size={22} /></Box>}
+        {!loadingPostcode ? (
+          <IconButton
+            aria-label='Search'
+            color='primary'
+            onClick={() => postcodeSearch()}
+            size="large">
+            <SearchIcon />
+          </IconButton>
+        ) : (
+          <Box
+            position='relative'
+            display='inline-flex'
+            sx={{
+              padding: theme => theme.spacing(1.5),
+              textAlign: 'center',
+              alignItems: 'center',
+              justifyContent: 'center',
+              verticalAlign: 'middle',
+              boxSizing: 'border-box'
+            }}
+          >
+            <CircularProgress size={22} />
+          </Box>
+        )}
       </Tooltip>
-      <Divider orientation='vertical' flexItem />
       <Tooltip title='Use your current location'>
-        {!loadingLocation
-          ? (
-            <IconButton
-              aria-label='Search by current location'
-              color='primary'
-              className={classes.iconButton}
-              onClick={() => getLocation()}
-            >
-              <MyLocationIcon />
-            </IconButton>
-          ) : <Box position='relative' display='inline-flex' className={classes.iconProgress}><CircularProgress size={22} /></Box>}
+        {!loadingLocation ? (
+          <IconButton
+            aria-label='Search by current location'
+            color='primary'
+            onClick={() => getLocation()}
+            size="large">
+            <MyLocationIcon />
+          </IconButton>
+        ) : (
+          <Box
+            position='relative'
+            display='inline-flex'
+          >
+            <CircularProgress size={22} />
+          </Box>
+        )}
       </Tooltip>
-      {settings
-        ? (
-          <Tooltip title='Change search settings'>
-            <IconButton
-              aria-label='Open search settings menu'
-              className={classes.iconButton}
-              color='secondary'
-              onClick={(e) => { openSettingsMenu(e) }}
-            >
-              <SettingsIcon />
-            </IconButton>
-          </Tooltip>
-        ) : null}
-      <Menu
-        id='mnu-settings'
-        anchorEl={anchor}
-        keepMounted
-        open={Boolean(anchor)}
-        onClose={() => closeSettingsMenu()}
-      >
-        <ListSubheader disableSticky>Search distance</ListSubheader>
-        <MenuItem onClick={() => setSearchDistance(1609)}>1 mile</MenuItem>
-        <MenuItem onClick={() => setSearchDistance(4827)}>3 miles</MenuItem>
-        <MenuItem onClick={() => setSearchDistance(8045)}>5 miles</MenuItem>
-        <MenuItem onClick={() => setSearchDistance(16090)}>10 miles</MenuItem>
-        <MenuItem onClick={() => setSearchDistance(32180)}>20 miles</MenuItem>
-        <MenuItem onClick={() => setSearchDistance(80450)}>50 miles</MenuItem>
-      </Menu>
-    </div>
-  )
+    </Box>
+  );
 }
 
-export default withRouter(PostcodeSearch)
+export default PostcodeSearch
