@@ -1,32 +1,18 @@
 import React, { useState } from 'react'
-import { withRouter } from 'react-router'
 
-import Button from '@material-ui/core/Button'
-import Chip from '@material-ui/core/Chip'
-import Menu from '@material-ui/core/Menu'
-import MenuItem from '@material-ui/core/MenuItem'
-import Tooltip from '@material-ui/core/Tooltip'
+import Button from '@mui/material/Button'
+import Chip from '@mui/material/Chip'
+import Menu from '@mui/material/Menu'
+import MenuItem from '@mui/material/MenuItem'
+import Tooltip from '@mui/material/Tooltip'
 
-import BusinessIcon from '@material-ui/icons/BusinessTwoTone'
-
-import { makeStyles } from '@material-ui/core/styles'
+import BusinessIcon from '@mui/icons-material/BusinessTwoTone'
 
 import { useApplicationStateValue } from './context/applicationState'
 import { useSearchStateValue } from './context/searchState'
 import { useViewStateValue } from './context/viewState'
 
-import * as urlHelper from './helpers/url'
-
-const useStyles = makeStyles((theme) => ({
-  button: {
-    marginRight: theme.spacing(1)
-  },
-  chip: {
-    marginRight: theme.spacing(1)
-  }
-}))
-
-function ServiceFilter (props) {
+function ServiceFilter() {
   const [{ services, serviceLookup }] = useApplicationStateValue()
   const [{ serviceFilter }, dispatchSearch] = useSearchStateValue()
   const [{ }, dispatchView] = useViewStateValue()//eslint-disable-line
@@ -38,29 +24,25 @@ function ServiceFilter (props) {
   const closeServiceMenu = () => setServiceMenuAnchor(null)
 
   const chooseService = (service) => {
+    closeServiceMenu()
     const coords = JSON.parse(service.bbox).coordinates[0]
     dispatchSearch({ type: 'FilterByService', service: service })
     dispatchView({ type: 'FitToBounds', bounds: [coords[0], coords[2]] })
-    urlHelper.addService(props.history, service.systemName)
-    closeServiceMenu()
   }
 
   const clearServiceFilter = () => {
     dispatchSearch({ type: 'ClearAll' })
-    urlHelper.clearService(props.history)
   }
-
-  const classes = useStyles()
 
   return (
     <>
       {serviceFilter.length === 0 ? (
         <Tooltip title='Choose library service'>
-          <Button color='primary' className={classes.button} onClick={(e) => openServiceMenu(e.currentTarget)} startIcon={<BusinessIcon />}>
-            Choose service
+          <Button color='primary' onClick={(e) => openServiceMenu(e.currentTarget)} startIcon={<BusinessIcon />}>
+            Choose library service
           </Button>
         </Tooltip>
-      ) : <Chip className={classes.chip} color='primary' onDelete={clearServiceFilter} label={serviceLookup[serviceFilter[0]].name} />}
+      ) : <Chip color='primary' onDelete={clearServiceFilter} label={serviceLookup[serviceFilter[0]].name} />}
       <Menu
         id='menu-library-service'
         anchorEl={serviceMenuAnchor}
@@ -80,4 +62,4 @@ function ServiceFilter (props) {
   )
 }
 
-export default withRouter(ServiceFilter)
+export default ServiceFilter
