@@ -5,7 +5,10 @@ import App from './App'
 
 import * as serviceWorker from './serviceWorker'
 
-import 'typeface-roboto'
+import '@fontsource/roboto/300.css'
+import '@fontsource/roboto/400.css'
+import '@fontsource/roboto/500.css'
+import '@fontsource/roboto/700.css'
 
 import 'maplibre-gl/dist/maplibre-gl.css'
 
