@@ -22,12 +22,17 @@ export class Service {
 export async function getServices () {
   const response = await axios.get(config.servicesApi)
   if (response && response.data && response.data.length > 0) {
-    return response.data.map(s => (new Service()).fromJson(s))
+    return response.data.map(s => new Service().fromJson(s))
   } else {
     return []
   }
 }
 
 export function getServiceSystemName (name) {
-  return name.split(', ').reverse().join(' ').replace(/[. ,:-]+/g, '-').toLowerCase()
+  return name
+    .split(', ')
+    .reverse()
+    .join(' ')
+    .replace(/[. ,:-]+/g, '-')
+    .toLowerCase()
 }
